Adding Bicep infra
Adding Bicep infra files for creating a Cosmos Free Tier instance and an
Azure Container App to host the API and site.

The API will now host the app inside the wwwroot folder as static
content. For now this can only be done by executing the build.ps1 script
in the scripts directory