Adding command handler for electrity purchases
One of the big gaps right now is that the average usage gets affected by
the previous month's closing amount. This means once a purchase is made,
the average usage goes way down as the difference is a massive negative
ammount. To combat this, I'm building out purchases separately. This
commit adds the handler capability to store purchases in their own
container.