Adding authorization for the API
The API now follows the BFF Auth pattern. This will ensure that the
endpoints are protected to only authenticated users.

The change involves adding Microsoft Identity with the Web App flow, a
reverse proxy using the YARP proxy and binding the Angular app to SSL
execution. Additionally, authentication will now happen with Cookies and
the OpenID Connect protocol for challenges.