Fix auth redirecting to HTTPs in Docker
Fix the issue where the sing-in redirect is pointint to http, because
Kestrel sees the current scheme as http when running in Docker in Azure.
Since Azure provides the HTTPs certificate traffic will be HTTPs, but
ASP.NET isn't aware.

To combat this, we now intercept the redirect URI and rewrite http:// to
https://