Use chart component and display loader
For the electricity usage component, use the new shared ChartComponent
from menlo-lib and remove own chart implementation.

This now enables us to add usage of the LoadingComponent to this screen
without affecting the rendering of the chart as the chart's element
bindings are now contained to itself so the Canvas doesn't need to be
zone aware.