Allow Authorization override
Allow authorization override for endpoints as authing on http calls is a
bit difficult right now. Need to think of a different way to test these.
Maybe there is another Entra app I can give the same permissions as a
user and then it can hit the endpoints at least?