Switch to ISO 8601 Duration for DateRangeService
The DateRangeService can be so much simpler now that we support
ISO 8601 Durations in the API. The old methods are subject to browser
date configuration. We still have the issue with the source date, but
for the relative durations we are not time zone dependent now.

The old methods on DateRangeService have been flagged as obsolete.