Update usage calc to account for missed days
The electricity usage calc now allows for missed days when calculating
usages. It now uses a queue as the main data-structure for itterating
over the usages. Therefore there is no longer a need to find the
previous day and usage is calculated as the difference between the
previous and current one.

The big caveat is that purchases do not follow this pattern, but in
reality, I won't be buying and not uploading on days that I don't take a
reading.