Add Money as a domain concept
Add features for representing money and a few currencies. It isn't meant
to be a complete list just yet, but to build out the conceptual model.
I looked at some libraries, but they all add things I'm not going to
need.

This includes a custom Newtonsoft.Json Converter for the currency.