Initial Aspire implementation
This doesn't fully work with the Cosmos VNEXT emulator so for now it is
expected that the local instance of the emulator will be run.

I'm still undecided on if I will change away from the IEvangelist
repository or if I'll wait for the emulator. Probably the prior as the
emulator is moving really slowly.